# api-docs/openapi.yaml
openapi: 3.1.0
info:
  title: Catalog BFF Service API
  version: 1.0.0
  description: |
    BFF que orquesta datos de PIM, Stock y Tenant services para el backoffice.
    
    ## Módulos
    - **Catálogo**: Lectura agregada de variantes con stock
    - **Backoffice**: CRUD de productos y variantes
    - **Inventario**: Listado y resumen orquestado (Stock + PIM)
    - **Admin**: Dashboard y métricas del marketplace
servers:
  - url: /
    description: Base URL
  - url: http://localhost:8085
    description: Servidor local de desarrollo
  - url: http://localhost:8001/catalog-bff
    description: A través de Kong API Gateway
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT para autenticación
    TenantId:
      type: apiKey
      in: header
      name: X-Tenant-ID
      description: UUID del tenant para operaciones multi-tenant
...
  schemas:
...
    InventoryItem:
      type: object
      description: Item de inventario (Stock + PIM enriquecido)
      properties:
        variant_sku:
          type: string
          description: SKU de la variante
        product_name:
          type: string
          description: Nombre del producto
        category_id:
          type: string
          format: uuid
          nullable: true
          description: ID de categoría (null si sin categoría)
        category_name:
          type: string
          description: Nombre de la categoría
        available_quantity:
          type: number
          format: double
          description: Cantidad disponible
        reserved_quantity:
          type: number
          format: double
          description: Cantidad reservada
        last_entry_at:
          type: string
          format: date-time
          nullable: true
          description: Fecha de última entrada
        sale_price:
          type: number
          format: double
          description: Precio de venta
        stock_value:
          type: number
          format: double
          description: Valor del stock (precio × cantidad)
    InventoryListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/InventoryItem'
        total_count:
          type: integer
          description: Total de items que coinciden con los filtros
        page:
          type: integer
          description: Página actual
        page_size:
          type: integer
          description: Tamaño de página
        total_pages:
          type: integer
          description: Total de páginas
    InventoryTotals:
      type: object
      properties:
        total_skus:
          type: integer
          description: Total de SKUs en inventario
        total_available:
          type: number
          format: double
          description: Suma de cantidades disponibles
        total_reserved:
          type: number
          format: double
          description: Suma de cantidades reservadas
        total_stock_value:
          type: number
          format: double
          description: Valor total del stock
    InventoryCategoryTotal:
      type: object
      properties:
        category_id:
          type: string
          format: uuid
          nullable: true
        category_name:
          type: string
        sku_count:
          type: integer
        available_quantity:
          type: number
          format: double
        reserved_quantity:
          type: number
          format: double
        stock_value:
          type: number
          format: double
    InventorySummaryResponse:
      type: object
      properties:
        totals:
          $ref: '#/components/schemas/InventoryTotals'
        by_category:
          type: array
          items:
            $ref: '#/components/schemas/InventoryCategoryTotal'

    # --- Catálogo ---
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Código de error
        message:
          type: string
          description: Mensaje legible
        details:
          type: object
          additionalProperties:
            type: string
paths:
...
  /api/v1/inventory:
    get:
      summary: Listar inventario
      description: Listado de inventario orquestado (Stock + PIM) con filtros y paginación
      operationId: listInventory
      tags: [Inventory]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Número de página
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            default: 20
          description: Tamaño de página
        - name: search
          in: query
          schema:
            type: string
          description: Búsqueda por nombre de producto o SKU
        - name: category_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filtrar por ID de categoría
        - name: min_available
          in: query
          schema:
            type: number
            format: double
          description: Cantidad mínima disponible
        - name: max_available
          in: query
          schema:
            type: number
            format: double
          description: Cantidad máxima disponible
        - name: min_price
          in: query
          schema:
            type: number
            format: double
          description: Precio mínimo
        - name: max_price
          in: query
          schema:
            type: number
            format: double
          description: Precio máximo
        - name: last_movement_from
          in: query
          schema:
            type: string
            format: date
            pattern: '^\d{4}-\d{2}-\d{2}$'
          description: Fecha desde (YYYY-MM-DD) última entrada
        - name: last_movement_to
          in: query
          schema:
            type: string
            format: date
            pattern: '^\d{4}-\d{2}-\d{2}$'
          description: Fecha hasta (YYYY-MM-DD) última entrada
        - name: sort_by
          in: query
          schema:
            type: string
            enum:
              - variant_sku
              - product_name
              - category_name
              - available_quantity
              - reserved_quantity
              - last_entry_at
              - sale_price
              - stock_value
            default: variant_sku
          description: Campo de ordenamiento
        - name: sort_dir
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
          description: Dirección de ordenamiento
      responses:
        '200':
          description: Lista de items de inventario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryListResponse'
        '400':
          description: X-Tenant-ID faltante o parámetros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error de orquestación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/inventory/summary:
    get:
      summary: Resumen de inventario
      description: Totalizador y desglose por categoría del inventario
      operationId: getInventorySummary
      tags: [Inventory]
      responses:
        '200':
          description: Resumen de inventario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventorySummaryResponse'
        '400':
          description: X-Tenant-ID faltante
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error de orquestación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...